Finish implementation of EchoClient.java
It seems like splitting the functionality of the program into two separate while loops was doing really terrible
things, with outputs seemingly just randomly stopping after a line or two.

Using a single while loop made this much easier (and it may have even been impossible with the two while loops)
to send a single byte at a time, as we could write it, send it, receive it, and write to the system all in the same loop.

Also added close() statements for the socket and it's input/output streams at the end.